Fix bugs

+++ subscriptions.py
@@ -1,4 +1,3 @@
-# coding: utf8
 import os.path
 from datetime import date
 from typing import NamedTuple
@@ -6,7 +5,7 @@ from typing import NamedTuple
 import telebot.types
 from tinkoff.invest import Operation
 
-from config import bot, REPORT_NAME, SUBSCRIPTION_MESSAGE_PATTERN, logger, BALANCE_SHORTCUT, ENCODING, RUBBLES_SHORTCUT
+from config import bot, REPORT_NAME, SUBSCRIPTION_MESSAGE_PATTERN, logger, BALANCE_SHORTCUT, RUBBLES_SHORTCUT
 from db import Database
 from exceptions import NotEnoughArguments, InvalidPortfolioID, InvalidTinkoffToken, InvalidNumber
 from tinkoffapi import TinkoffApi
@@ -36,36 +35,56 @@ class UnsubscriptionMessage(NamedTuple):
 
 class ReportUnit:
     """Структура составной части отчёта"""
-    figi: str
-    name: str
-    ticker: str
-    currency: str
-    balance: float
-    bought_at_sum: int
-    fee: float
-    profit: Profit
+    _figi: str
+    _name: str
+    _ticker: str
+    _currency: str
+    _balance: float
+    _bought_at_sum: int
+    _fee: int
+    _absolute_profit: Profit or str
+    _relative_profit: Profit
 
     def __init__(self, *args):
-        self.figi = args[0]
-        self.name = args[1]
-        self.ticker = args[2]
-        self.currency = args[3]
-        self.balance = args[4]
-        self.bought_at_sum = args[5]
-        self.fee = args[6]
-        self.profit = args[7]
+        self._figi = args[0]
+        self._name = args[1]
+        self._ticker = args[2]
+        self._currency = args[3]
+        self._balance = args[4]
+        self._bought_at_sum = args[5]
+        self._fee = args[6]
+        self._absolute_profit = args[7]
+        self._relative_profit = args[8]
 
-    def set_profit(self, profit: Profit):
-        self.profit = profit
+    def set_relative_profit(self, relative_profit: Profit):
+        self._relative_profit = relative_profit
+
+    def get_figi(self):
+        return self._figi
+
+    def get_name(self):
+        return self._name
+
+    def get_ticker(self):
+        return self._ticker
+
+    def get_currency(self):
+        return self._currency
 
     def get_balance(self):
-        return self.balance
+        return self._balance
 
     def get_bought_at_sum(self):
-        return self.bought_at_sum
+        return self._bought_at_sum
 
     def get_fee(self):
-        return self.fee
+        return self._fee
+
+    def get_absolute_profit(self):
+        return self._absolute_profit
+
+    def get_relative_profit(self):
+        return self._relative_profit
 
 
 @handler
@@ -144,15 +163,19 @@ def _parse_unsubscription_message(msg: telebot.types.Message) \
     return UnsubscriptionMessage(broker_account_id)
 
 
+@handler
 def _notify(api: TinkoffApi, user_id: int):
     try:
         _form_report(api)
         with open(REPORT_NAME, "rb") as f:
             bot.send_document(user_id, f)
     except Exception as e:
-        bot.send_message(user_id, e)
+        bot.send_message(user_id, "Произошла фатальная ошибка! Пожалуйста, сообщите о ней администратору.")
+        print(e)
+        logger.error(e)
     finally:
-        os.remove(REPORT_NAME)
+        if os.path.exists(os.path.realpath(REPORT_NAME)):
+            os.remove(REPORT_NAME)
 
 
 def _parse_api(raw_api: tuple) \
@@ -169,7 +192,7 @@ def _form_report(api: TinkoffApi):
     csv_rows = _get_csv_rows(operations_map, api)
     csv_rows = list_to_string(csv_rows)
 
-    with open(REPORT_NAME, "w", encoding=ENCODING) as f:
+    with open(REPORT_NAME, "w") as f:
         f.write(csv_rows)
 
 
@@ -202,7 +225,8 @@ def _get_operations_map(api: TinkoffApi) \
                 bought_at_sum = -int(get_canonical_price(operation.payment))
                 balance = -operation.quantity
 
-        profit = NULL_PROFIT
+        absolute_profit = NULL_PROFIT
+        relative_profit = NULL_PROFIT
 
         if _is_usd(operation):
             usd_course = api.get_usd_course()
@@ -214,19 +238,44 @@ def _get_operations_map(api: TinkoffApi) \
             bought_at_sum += res[name].get_bought_at_sum()
             fee += res[name].get_fee()
 
-        report_unit = ReportUnit(figi, name, ticker, currency, balance, bought_at_sum, fee, profit)
+        report_unit\
+            = ReportUnit(figi, name, ticker, currency, balance, bought_at_sum, fee, absolute_profit, relative_profit)
         res[name] = report_unit
 
     for position in res.values():
-        profit = _get_profit(position, api)
-        position.set_profit(profit)
+        relative_profit = _get_relative_profit(position, api)
+        position.set_relative_profit(relative_profit)
 
     return res, total_inputs_sum
 
 
 def _get_csv_rows(operations_map: tuple[dict[str, ReportUnit], int], api: TinkoffApi) \
         -> list[str]:
-    csv_rows = [",".join([
+    csv_rows = _form_csv_titles()
+
+    total_bought_at_sum = 0
+    total_fee_sum = 0
+    total_portfolio_sum = 0
+
+    for report_unit in operations_map[0].values():
+        completed_csv_row = _get_completed_csv_row(report_unit)
+        _append_csv_row(completed_csv_row, csv_rows)
+
+        total_bought_at_sum += report_unit.get_bought_at_sum()
+        total_fee_sum += report_unit.get_fee()
+        total_portfolio_sum += int(report_unit.get_balance() * api.get_price(report_unit.get_figi()))
+
+    total_inputs_sum = operations_map[1]
+    total_completed_csv_row = _get_total_completed_csv_row(total_bought_at_sum, total_fee_sum,
+                                                           total_portfolio_sum, total_inputs_sum)
+    _append_csv_row(total_completed_csv_row, csv_rows)
+
+    return csv_rows
+
+
+def _form_csv_titles() \
+        -> list[str]:
+    return [",".join([
         "securities name",
         "ticker",
         "currency",
@@ -237,38 +286,38 @@ def _get_csv_rows(operations_map: tuple[dict[str, ReportUnit], int], api: Tinkof
         "relative profit"
     ])]
 
-    total_bought_at_sum = 0
-    total_fee_sum = 0
-    total_portfolio_sum = 0
 
-    for report_unit in operations_map[0].values():
-        name = report_unit.name
-        ticker = report_unit.ticker
-        currency = report_unit.currency
-        balance = str(report_unit.balance) + " " + BALANCE_SHORTCUT
-        bought_at_sum = str(report_unit.bought_at_sum)
-        fee = to_rub(report_unit.fee)
-        absolute_profit = to_rub(report_unit.profit.absolute)
-        relative_profit = int(report_unit.profit.relative)
-        relative_profit = f"{absolute_profit} ({relative_profit}%)"
-        absolute_profit = "-"
+def _get_completed_csv_row(report_unit: ReportUnit) \
+        -> ReportUnit:
+    figi = report_unit.get_figi()
+    name = report_unit.get_name()
+    ticker = report_unit.get_ticker()
+    currency = report_unit.get_currency()
+    balance = str(report_unit.get_balance()) + " " + BALANCE_SHORTCUT
+    bought_at_sum = str(report_unit.get_bought_at_sum())
+    fee = to_rub(report_unit.get_fee())
+    absolute_profit = to_rub(report_unit.get_relative_profit().absolute)
+    relative_profit = int(report_unit.get_relative_profit().relative)
+    relative_profit = f"{absolute_profit} ({relative_profit}%)"
+    absolute_profit = "-"
 
-        total_bought_at_sum += report_unit.bought_at_sum
-        total_fee_sum += report_unit.fee
-        total_portfolio_sum += int(report_unit.balance * api.get_price(report_unit.figi))
+    return ReportUnit(
+        figi,
+        name,
+        ticker,
+        currency,
+        balance,
+        bought_at_sum,
+        fee,
+        absolute_profit,
+        relative_profit
+    )
 
-        csv_rows.append(",".join([
-            name,
-            ticker,
-            currency,
-            balance,
-            bought_at_sum,
-            fee,
-            absolute_profit,
-            relative_profit
-        ]))
 
-    total_inputs_sum = operations_map[1]
+def _get_total_completed_csv_row(total_bought_at_sum: int, total_fee_sum: int,
+                                 total_portfolio_sum: int, total_inputs_sum: int) \
+        -> ReportUnit:
+    figi = ""
     total_relative_absolute_profit = total_portfolio_sum - total_bought_at_sum
     total_relative_relative_profit = int(
         100.0 * total_relative_absolute_profit / total_bought_at_sum) if total_bought_at_sum != 0 else 0
@@ -284,7 +333,8 @@ def _get_csv_rows(operations_map: tuple[dict[str, ReportUnit], int], api: Tinkof
     total_bought_at_sum = str(total_bought_at_sum)
     total_fee_sum = to_rub(total_fee_sum)
 
-    csv_rows.append(",".join([
+    return ReportUnit(
+        figi,
         "Total",
         "-",
         RUBBLES_SHORTCUT,
@@ -293,15 +343,26 @@ def _get_csv_rows(operations_map: tuple[dict[str, ReportUnit], int], api: Tinkof
         total_fee_sum,
         total_absolute_profit,
         total_relative_profit
+    )
+
+
+def _append_csv_row(csv_row: ReportUnit, csv_rows: list[str]):
+    csv_rows.append(",".join([
+        csv_row.get_name(),
+        csv_row.get_ticker(),
+        csv_row.get_currency(),
+        csv_row.get_balance(),
+        csv_row.get_bought_at_sum(),
+        csv_row.get_fee(),
+        csv_row.get_absolute_profit(),
+        csv_row.get_relative_profit()
     ]))
 
-    return csv_rows
 
-
-def _get_profit(report_unit: ReportUnit, api: TinkoffApi) \
+def _get_relative_profit(report_unit: ReportUnit, api: TinkoffApi) \
         -> Profit:
-    bought_at_sum = report_unit.bought_at_sum
-    cur_asset_value = report_unit.balance * api.get_price(report_unit.figi)
+    bought_at_sum = report_unit.get_bought_at_sum()
+    cur_asset_value = report_unit.get_balance() * api.get_price(report_unit.get_figi())
     absolute_profit = int(cur_asset_value - bought_at_sum)
     relative_profit = int(100 * absolute_profit / bought_at_sum) if bought_at_sum != 0 else 0
     return Profit(absolute_profit, relative_profit)
